🚚 Rename: Reame the folder from `supermarket_problem` to `supermarket_implementation`
